use coord. xact for truncate commands as we do for drop commands
Before this commit, we might hit to assertion crash in local
executor due to following check:
  isValidLocalExecutionPath |= !AnyTaskAccessesLocalNode(taskList);

This is because, for truncate commands, we create task lists for
relations one by one as we execute truncate commands on shards
via per-relation truncate triggers. However, local executor expects
local tasks all at once and we might hit to below assertion.
  Assert(isValidLocalExecutionPath);

This was not the case for reference tables as we already use 2PC when
executing truncate commands on reference tables.

A better solution would be to create task list for all relations at
once for both truncate and drop commands and pass the tasks to
related triggers somehow and execute those tasklists when the trigger
is executed for the first relation that postgres decides to drop or
truncate.