Explain why VacuumStmtParams is a copy of ExecVacuum
It seems that the reason VacuumStmtParams is almost a copy of ExecVacuum
on postgres side is because ExecVacuum also does the vacuum at the end.
We don't want to do the vacuum in the coordinator hence we copied this
method. However if postgres separated this method into two methods to
first check if we need to error and then do the actual vacuum, we could
just use the same method.